pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "toughmistyloaderprogram"
version = "0.1.0"
[tool.poetry.dependencies]
PyNaCl = "^1.3"
discord = "^1.0"
python = "^3.8"
giphy_client = "^1.0.0"
python-dotenv = "^0.21.0"
Flask = "^2.2.2"
yt-dlp = "^2023.2.17"
youtube_dl = "^2021.12.17"
[tool.poetry.dev-dependencies]
